Addtion of Dictionary and Excel class
Dictionary :
This is the model that will used as a basis for the entire game.

Attributes :
words, themes, definitions (dictionaries), fileName, sheetName (for Excel), size

Methods :
fill, put, delete (not implemented yet), getSheet, getFile, getSize, main (for testing the class)

Excel :
This is a template that can import, modify and save an Excel file.

All methods has been testing.